% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predic(t)amentSolution.R
\name{ExportRPFVals}
\alias{ExportRPFVals}
\title{Export R values, P values and F values for each column into a CSV}
\usage{
ExportRPFVals()
}
\arguments{
\item{NA}{no parameters needed}
}
\value{
No value returned
}
\description{
This functionCalculates R values, P values and F values for each column and export these into a CSV
to be able to compare which columns affect the sale price the most to then be used in the prediction
}
