% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predic(t)amentSolution.R
\name{FeatureScalling}
\alias{FeatureScalling}
\title{Feature Scalling function}
\usage{
FeatureScalling(x)
}
\arguments{
\item{DatasetColumn}{Takes in a column using lapply}
}
\value{
List of numeric values between 0-1
}
\description{
This function takes in string or numeric values and formats them into a number between 0 and 1
}
