% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predic(t)amentSolution.R
\name{importLibrariesAndDB}
\alias{importLibrariesAndDB}
\title{Install and import necessary libraries and import dataset from SQL}
\usage{
importLibrariesAndDB()
}
\arguments{
\item{NA}{no parameters needed}
}
\value{
The train dataset
}
\description{
sample description
sample2
}
\examples{
trainData <- importLibrariesAndDB()
}
