% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{normaliseData}
\alias{normaliseData}
\title{Normalise train dataset to use only relevant columns and rows}
\usage{
normaliseData(dat)
}
\arguments{
\item{Dataset}{Takes in the train dataset}
}
\value{
Normalised dataset
}
\description{
This function creates a subset of the data frame with pre-selected columns and removes rows with
NA values
}
\examples{
dataNormalised <- normaliseData(trainData)
}
