% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predic(t)amentSolution.R
\name{weightedKNNModel}
\alias{weightedKNNModel}
\title{Weighted K-Nearest Neighbor model}
\usage{
weightedKNNModel()
}
\arguments{
\item{NA}{no parameters needed}
}
\value{
Weighted K-Nearest Neighbor model
}
\description{
This function calculates a k value and trains a weighted knn model with the train data
}
