% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predic(t)amentSolution.R
\name{weightedKNNPredict}
\alias{weightedKNNPredict}
\title{Weighted K-Nearest Neighbor prediction}
\usage{
weightedKNNPredict(input)
}
\arguments{
\item{list}{List with user inputs from UI}
}
\value{
Predicted sale price rounded to the nearest thousand
}
\description{
This function calculates a sale price prediction based on a previously trained K-Nearest Neighbormodel and
the user input
}
