% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predic(t)amentSolution.R
\name{importLibrariesAndDB}
\alias{importLibrariesAndDB}
\title{Install and import necessary libraries and import dataset from SQL}
\usage{
importLibrariesAndDB()
}
\arguments{
\item{NA}{no parameters needed}
}
\value{
The training dataset
}
\description{
This function contains a list of necessary libraries for the application to run
it checks if the libraries are already installed, if they arents it proceeds to installation.
Then it imports each library, connects to a local database to extract the data for the data set
and returns the dataset with the train data.
}
\examples{
trainData <- importLibrariesAndDB()
}
